+++ civis/ml/tests/test_helper.py
@@ -1,7 +1,10 @@
 import pytest
 
+from civis.compat import mock
 from civis.response import Response
-from civis.ml import list_models
+from civis.ml import (list_models, put_models_shares_groups,
+                      put_models_shares_users)
+from civis.ml import _helper as helper
 from civis.tests.mocks import create_client_mock
 
 
@@ -22,3 +25,94 @@ def test_list_models():
 
     out = list_models(job_type=None, client=m_client)
     assert out == resp
+
+
+def _create_share_model_client_mock(run_ids):
+    m_client = create_client_mock()
+    m_client.scripts.put_containers_shares_users.return_value = 'usershare'
+    m_client.scripts.put_containers_shares_groups.return_value = 'groupshare'
+    m_client.scripts.list_containers_runs.return_value = [
+        Response({'id': _id}) for _id in run_ids]
+    m_client.scripts.list_containers_runs_outputs.return_value = [
+        Response({'object_id': 117, 'object_type': 'File'}),
+        Response({'object_id': 31, 'object_type': 'Project'}),
+        Response({'object_id': 37, 'object_type': 'JSONValue'}),
+    ]
+    return m_client
+
+
+def test_share_model_users():
+    m_client = _create_share_model_client_mock([11])
+
+    resp = helper._share_model(3, [7, 8], 'write', 'users', client=m_client,
+                               send_shared_email=True)
+    assert resp == 'usershare'
+    m_client.scripts.put_containers_shares_users.assert_called_once_with(
+        3, [7, 8], 'write', send_shared_email=True)
+    m_client.files.put_shares_users.assert_called_once_with(
+        117, [7, 8], 'write', send_shared_email=False)
+    m_client.projects.put_shares_users.assert_called_once_with(
+        31, [7, 8], 'write', send_shared_email=False)
+    m_client.json_values.put_shares_users.assert_called_once_with(
+        37, [7, 8], 'write', send_shared_email=False)
+
+
+def test_share_model_groups():
+    m_client = _create_share_model_client_mock([11])
+
+    resp = helper._share_model(3, [7, 8], 'write', 'groups', client=m_client,
+                               send_shared_email=True)
+    assert resp == 'groupshare'
+    m_client.scripts.put_containers_shares_groups.assert_called_once_with(
+        3, [7, 8], 'write', send_shared_email=True)
+    m_client.files.put_shares_groups.assert_called_once_with(
+        117, [7, 8], 'write', send_shared_email=False)
+    m_client.projects.put_shares_groups.assert_called_once_with(
+        31, [7, 8], 'write', send_shared_email=False)
+    m_client.json_values.put_shares_groups.assert_called_once_with(
+        37, [7, 8], 'write', send_shared_email=False)
+
+
+def test_share_model_tworuns():
+    # Check that we grant permission on run outputs for each run
+    m_client = _create_share_model_client_mock([11, 13])
+
+    helper._share_model(3, [7, 8], 'write', 'users', client=m_client)
+
+    m_client.scripts.put_containers_shares_users.assert_called_once_with(
+        3, [7, 8], 'write')
+
+    assert m_client.files.put_shares_users.call_count == 2
+    assert m_client.projects.put_shares_users.call_count == 2
+    assert m_client.json_values.put_shares_users.call_count == 2
+
+
+def test_share_model_project_permissions():
+    # Grant "write" permission on the internal project when
+    # overall "read" permission is requested.
+    m_client = _create_share_model_client_mock([11])
+
+    helper._share_model(3, [7, 8], 'read', 'groups', client=m_client)
+
+    m_client.projects.put_shares_groups.assert_called_once_with(
+        31, [7, 8], 'write', send_shared_email=False)
+
+
+@mock.patch('civis.ml._helper._share_model')
+def test_put_models_shares_groups(mock_share):
+    mock_share.return_value = 'retval'
+    out = put_models_shares_groups(1, [7, 8], 'read')
+
+    assert out == 'retval'
+    mock_share.assert_called_once_with(1, [7, 8], 'read', entity_type='groups',
+                                       client=None)
+
+
+@mock.patch('civis.ml._helper._share_model')
+def test_put_models_shares_groups(mock_share):
+    mock_share.return_value = 'retval'
+    out = put_models_shares_users(1, [7, 8], 'read')
+
+    assert out == 'retval'
+    mock_share.assert_called_once_with(1, [7, 8], 'read', entity_type='users',
+                                       client=None)